结算保留一位数字

+++ lib/components/settle.dart
@@ -30,7 +30,7 @@ class Settle extends StatelessWidget {
               mainAxisSize: MainAxisSize.min,
               children: realSettle.asMap().entries.map((entry) {
                 int index = entry.key;
-                int score = entry.value;
+                double score = entry.value;
                 return ListTile(
                   title: Text('${playController.names[index]} : ${score} 元'),
                 );
